added a script for video synchronization that implements the Average Hash algorithm.
Added in some tests that check the execution time for both methods of synchronization.

Added in logs for the tests that show individual execution times and average execution times.

Created a To Do text file to keep everyone on the same page of what needs to be done.

Started working on a DCT pHash for video synchronization. It is a WIP as it currently doesn't work.