Added a video synchronization method that utilizes Simple Subtraction and tested the execution time.
The test I ran ran each video synchronization method 1000 times and logged the results in tests/logs. The average execution time can be found at the bottom of each log.

The average hash gave the best result in this particular test, but we may want to run more tests and also design better tests to ensure this is consistently the most efficient method of video synchronization.