Improved pixelwise thresholds. Included very basic tests and logs of results. Included a demo file for easy demonstration and learning.
In pixelwise_video_sync.py:
Changed difference threshold from 200 to 50 so smaller differences are found.
Changed matching threshold to 20 so that only very similar frames are matched, with a small margin of error.

Demo1 is a video player with no synchronization.
Demo2 is the pixelwise video sync, but it shows every step the program takes to synchronize the videos.
Demo3 is identical to pixelwise_video_sync.py but has higher thresholds.

f_pixelwise_video_sync.py is a faster version of pixelwise_video_sync.py because it converts some frames to grayscale outside loops.

tests/logic_only contains scripts that use different methods for video synchronization. These specific scripts don't show any windows and don't print anything out. They are solely for testing execution time.

tests/logs contains logs of the execution time of the tests in tests/logic_only. The average execution time is logged at the bottom line of each file.

optical_flow_analysis detects object movement in the video. It doesn't seem great for our use case due to the simplicity of our videos, but we can play around with it. We may be able to increase its sensitivity and use the annotations it makes to check for differences in each video and create bounding boxes or something.